requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart
alembic
azure-storage-blob
dotenv
openai
pgvector
rembg
pillow
onnxruntime